Update costFunctionReg.m [Formula for J]
[On line number: 25]
The current formula to find the cost (J) fails, as it also includes the theta(1) parameter in the calculations. We should not be regularizing the theta(1) parameter in the code as theta(1) parameter is not affected by the lambda parameter in logistic regression. Hence, that theta(1) parameter must be removed. So, I used "theta(2:length(theta))" instead of "theta" in the formula to find J, so as to exclude theta(1) parameter and find the correct value for J.